Made Jumper to avoid remote require
