+++ main.js
@@ -84,7 +84,7 @@ app.on('ready', () => {
 
   // Load IPC handler
   require(utilsPath + '/ipcHandler')
-  require(utilsPath + '/interactive')
+  require(utilsPath + '/jumper')
 
   mainWindow.on('closed', () => {
     mainWindow = null
